requirements.txt
facebook-scraper==0.2.17
requests==2.22.0
bs4==0.0.1
readability==0.3.1
boto3==1.16.28
pandas==1.2.0
python-decouple==3.4